feat: move remaining collection engine modules into the package (chunk 1c-iv/v)
Move the last engine modules — derive.ts (enrichItems), spawn.ts (recurrence /
successor computation), delete.ts (delete + archive), views.ts (custom-view
CRUD) — into @mulmoclaude/collection-plugin/server, rebinding their host deps
via the binding: log → host proxy, workspacePath → getWorkspaceRoot(),
WORKSPACE_DIRS.{skillsStaging,archive} + isPresetSlug → the binding's path
helpers, writeFileAtomic → the package port, errorMessage/ONE_DAY_MS → a small
package util.

The host-integration modules stay host-side and call the package: notifications.ts
(host notifier engine), watcher.ts (fs.watch + host pubsub), api/routes/
collections.ts (express), manageCollection.ts (MCP tool) — all reach the engine
through the collections barrel / package server entry.

The collection engine (schema, derive/formula/visibility/sort/calendar, storage,
validation, discovery, delete, views, spawn) now lives entirely in the package,
parameterized over the host via configureCollectionHost. Phase 1 server-side
extraction complete.

Verified: yarn typecheck (server/vue/test/packages), lint, build all green;
full unit suite (5102 tests) 0 failures.
