+++ packages/plugins/collection-plugin/src/server/index.ts
@@ -13,3 +13,7 @@ export * from "./templatePath";
 export * from "./io";
 export * from "./validate";
 export * from "./discovery";
+export * from "./derive";
+export * from "./spawn";
+export * from "./delete";
+export * from "./views";
