+++ server/workspace/collections/index.ts
@@ -1,8 +1,8 @@
 export { discoverCollections, loadCollection, toSummary, toDetail, CollectionSchemaZ, type LoadedCollection } from "@mulmoclaude/collection-plugin/server";
 export { validateCollectionRecords, validateRecordObject, COMPUTED_TYPES, type RecordIssue } from "@mulmoclaude/collection-plugin/server";
-export { enrichItems } from "./derive.js";
-export { deleteCollection, deleteCollectionRefusalMessage, type DeleteCollectionResult } from "./delete.js";
-export { deleteCustomView, type DeleteViewResult } from "./views.js";
+export { enrichItems } from "@mulmoclaude/collection-plugin/server";
+export { deleteCollection, deleteCollectionRefusalMessage, type DeleteCollectionResult } from "@mulmoclaude/collection-plugin/server";
+export { deleteCustomView, type DeleteViewResult } from "@mulmoclaude/collection-plugin/server";
 export {
   listItems,
   readItem,
