+++ server/workspace/collections/notifications.ts
@@ -42,8 +42,15 @@ import type { NotifierEntry } from "../../notifier/types.js";
 import { log } from "../../system/logger/index.js";
 import { errorMessage } from "../../utils/errors.js";
 import { whenMatches } from "@mulmoclaude/collection-plugin";
-import { loadCollection, type DiscoveryOptions, listItems, readItem, type IoOptions } from "@mulmoclaude/collection-plugin/server";
-import { isTriggerDue, maybeSpawnSuccessor } from "./spawn.js";
+import {
+  loadCollection,
+  type DiscoveryOptions,
+  listItems,
+  readItem,
+  type IoOptions,
+  isTriggerDue,
+  maybeSpawnSuccessor,
+} from "@mulmoclaude/collection-plugin/server";
 import type { CollectionItem, CollectionSchema } from "./types.js";
 
 /** The legacy-id prefix every collection-completion bell entry carries.
