+++ test/workspace/collections/test_spawn.ts
@@ -20,8 +20,9 @@ import {
   parseCivil,
   successorId,
   type CivilDate,
-} from "../../../server/workspace/collections/spawn.js";
-import { readItem, writeItem } from "@mulmoclaude/collection-plugin/server";
+  readItem,
+  writeItem,
+} from "@mulmoclaude/collection-plugin/server";
 import type { CollectionEvery, CollectionSchema } from "../../../server/workspace/collections/types.js";
 
 describe("daysInMonth", () => {
